// microservicios-web/src/app/views/dashboard/Setup/venta/models/venta.ts
export class Venta {
  id?: number;
  fechaVenta?: string;
  vendedor?: string;
  numeroSerie?: string;
  modeloCelular?: string;
  marca?: string;
  precioVenta?: number;
  metodoPago?: string;
  estadoCelular?: string;
  condicionCelular?: string;
  accesoriosIncluidos?: string;
  garantia?: string;
  ubicacionVenta?: string;
  notas?: string;
  estadoVenta?: string;
  canalesVenta?: string;
  descuentosPromociones?: number;
  costosEnvio?: number;
  informacionEnvio?: string;

}